fix: split the data logic
